Refactor code and add XO
