+++ .gitignore
@@ -2,3 +2,4 @@ yarn.lock
 node_modules
 .next
 *.log*
+.env
