+++ client/package.json
@@ -1,5 +1,5 @@
 {
-	"name": "knowledge",
+	"name": "knowledge-client",
 	"version": "0.0.0",
 	"private": true,
 	"engines": {
