+++ server/package.json
@@ -1,0 +1,31 @@
+{
+	"name": "knowledge-server",
+	"version": "0.0.0",
+	"private": true,
+	"main": "index.js",
+	"engines": {
+		"node": ">=8"
+	},
+	"scripts": {
+		"start": "node index.js",
+		"dev": "nodemon index.js",
+		"lint": "xo",
+		"test": "npm run lint"
+	},
+	"dependencies": {
+		"express": "^4.16.4",
+		"mongoose": "^5.3.14"
+	},
+	"devDependencies": {
+		"nodemon": "^1.18.7",
+		"xo": "^0.23.0"
+	},
+	"xo": {
+		"rules": {
+			"quotes": [
+				"error",
+				"double"
+			]
+		}
+	}
+}
