+++ server/readme.md
@@ -1,0 +1,6 @@
+## Getting started
+
+- Run `npm install` to install all of the required dependencies.
+- Run `docker-compose up` to build the Docker image of the server and run it, along with the MongoDB container.
+
+The test route can be at http://localhost:5000/user-get-all.
