--- server_nosub/.dockerignore
@@ -1,2 +1,0 @@
-/node_modules
-.git
