--- server_nosub/components/users/userRoutes.js
@@ -1,8 +1,0 @@
-const express = require("express");
-const router = express.Router();
-
-const { userGetAll } = require("./userControllers");
-
-router.get("/user-get-all", userGetAll);
-
-module.exports = router;
