--- server_nosub/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "server",
-  "version": "1.0.0",
-  "description": "Server built for q&a app",
-  "main": "server.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "nodemon server.js"
-  },
-  "author": "bears-team-22",
-  "license": "ISC",
-  "dependencies": {
-    "dotenv": "^6.1.0",
-    "express": "^4.16.4",
-    "mongoose": "^5.3.14"
-  },
-  "devDependencies": {
-    "nodemon": "^1.18.7"
-  }
-}
